+++ examples/richeditor/src/components/MyEditor.vue
@@ -2,13 +2,16 @@
 import { TiptapRichEditor } from 'tiptap-richeditorv-vue3'
 import UploaderButton from './UploaderButton.vue'
 import 'tiptap-richeditorv-vue3/style.css'
+import { Form as TForm, FormItem as TFormItem, Input as TInput, Textarea as TTextarea } from 'tdesign-vue-next'
+import 'tdesign-vue-next/es/style/index.css'
 import { ref } from 'vue';
+console.log(TInput)
 const content = ref('<p>h2</p><img src="https://public.app.wfell.top/storage/v1/render/image/public/demo/Case%20Study/PJ1018-AerospaceBlade/Aerospace%20Blade1.png?width=600" />')
 </script>
 
 <template>
   <div class="card">
-    <TiptapRichEditor v-model="content" :uiuploader="UploaderButton" />
+    <TiptapRichEditor v-model="content" :uiinput="TInput" :uitextarea="TTextarea" :uiuploader="UploaderButton" :uiform="TForm" :uiformitem="TFormItem" />
   </div>
   <div class="card">
     <h3>html输出</h3>